Optimization of news app.
