+++ news/views/__init__.py
@@ -1,0 +1,4 @@
+from .common import *
+from .coins import *
+from .posts import *
+from .tags import *
